Refactor theme switch into its own full stack component
