+++ app/routes/_auth+/login.tsx
@@ -9,6 +9,7 @@ import { Form, NavLink, useActionData } from '@remix-run/react';
 import { z } from 'zod';
 import { Field } from '~/app/components/forms';
 import { Button } from '~/app/components/ui/button';
+import ThemeSwitch from '~/app/routes/_resources+/theme-switch';
 import { createLoginSession, requireAnonymus } from '~/app/utils/auth.server';
 import { EmailSchema, RememberSchema } from '~/app/utils/validation-schemas';
 import { handleNewSession } from './login.server';
@@ -77,6 +78,7 @@ export default function LoginRoute() {
 	return (
 		<div className="flex min-h-screen flex-col lg:p-8 lg:pb-16">
 			<div className="flex">
+				<ThemeSwitch />
 				<Button asChild className="ml-auto" variant="ghost">
 					<NavLink to="/signup">Signup</NavLink>
 				</Button>
